seperate test_backgroundsubtraction from test_movementdetection
